+++ Common/src/main/java/de/markusbordihn/easymodelentities/network/animation/EasyModelAnimationNetwork.java
@@ -26,7 +26,7 @@ import net.minecraft.world.entity.Entity;
 
 public final class EasyModelAnimationNetwork {
 
-  public static final int PROTOCOL_VERSION = 1;
+  public static final int PROTOCOL_VERSION = 2;
 
   private static volatile Sender sender = Sender.NONE;
 
@@ -45,6 +45,15 @@ public final class EasyModelAnimationNetwork {
     sender.send(level, blockPos, packet);
   }
 
+  public static void send(Entity entity, ClientboundEasyModelAnimationSequencePacket packet) {
+    sender.send(entity, packet);
+  }
+
+  public static void send(
+      ServerLevel level, BlockPos blockPos, ClientboundEasyModelAnimationSequencePacket packet) {
+    sender.send(level, blockPos, packet);
+  }
+
   public interface Sender {
 
     Sender NONE = new Sender() {};
@@ -53,5 +62,10 @@ public final class EasyModelAnimationNetwork {
 
     default void send(
         ServerLevel level, BlockPos blockPos, ClientboundEasyModelAnimationPacket packet) {}
+
+    default void send(Entity entity, ClientboundEasyModelAnimationSequencePacket packet) {}
+
+    default void send(
+        ServerLevel level, BlockPos blockPos, ClientboundEasyModelAnimationSequencePacket packet) {}
   }
 }